LayoutTests/webaudio/resume-context-while-running-expected.txt

Tests that calling AudioContext.resume() while already running resolves the promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Did start rendering
PASS context.state is "running"
PASS Calling resume() while running resolved the promise
PASS context.state is "running"
PASS Calling resume() while running resolved the promise
PASS context.state is "running"
PASS successfullyParsed is true

TEST COMPLETE
